+++ Titres/Outils.py
@@ -53,59 +53,38 @@ def gainCoins(user,coins):
     curseur.execute("UPDATE coins SET Coins=Coins+{0}".format(coins))
     connexion.commit()
 
-def dailyCoins():
-    dictRank={5:2,4:4,3:6,2:8,1:10}
-    liste=["P4","BatailleNavale","Tortues","TortuesDuo","TrivialVersus","TrivialParty","TrivialBR"]
-    for i in liste:
-        connexion,curseur=connectSQL("OT",i,"Jeux","GL","")
-        for j in curseur.execute("SELECT * FROM glob WHERE Rank<=5").fetchall():
-            gainCoins(j["ID"],dictRank[j["Rank"]]) 
+def titresTrivial(niveau,categ,user):
+    connexionUser,curseurUser=connectSQL("OT",user,"Titres",None,None)
+    if niveau==5:
+        dict5={0:"Cultivé",1:"Diverti",2:"Scientifique",3:"Mythe",4:"Sportif",5:"Géographe",6:"Historien",7:"Député",8:"Artiste",9:"Célèbre",10:"Lion",11:"Mécano",12:"Cerveau"}
+        dict5ID={0:162,1:11,2:13,3:15,4:17,5:19,6:21,7:23,8:25,9:27,10:29,11:31,12:9}
+        curseurUser.execute("INSERT INTO titresUser VALUES({0},'{1}',0)".format(dict5ID[categ],dict5[categ]))
+    else:
+        dict10={0:"Connaisseur",1:"Réalisateur",2:"Virologue",3:"Légende",4:"Athlète",5:"Cartographe",6:"Historique",7:"Ministre",8:"Chef d’Oeuvre",9:"Star",10:"Requin",11:"Bolide",12:"Encyclopédie"}
+        dict10ID={0:163,1:12,2:14,3:16,4:18,5:20,6:22,7:24,8:26,9:28,10:30,11:32,12:10}
+        curseurUser.execute("INSERT INTO titresUser VALUES({0},'{1}',0)".format(dict10ID[categ],dict10[categ]))
+    connexionUser.commit()
 
+def titresJeux(wins,option,user):
+    connexionUser,curseurUser=connectSQL("OT",user,"Titres",None,None)
+    if wins==5:
+        dict5={"P4":"Aligné","BatailleNavale":"Navire","Tortues":"Empilé","TortuesDuo":"Équipier","TrivialVersus":"Vainqueur","TrivialParty":"Fêtard","TrivialBR":"Vivant"}
+        dict5ID={"P4":39,"BatailleNavale":45,"Tortues":41,"TortuesDuo":43,"TrivialVersus":33,"TrivialParty":35,"TrivialBR":37}
+        curseurUser.execute("INSERT INTO titresUser VALUES({0},'{1}',0)".format(dict5ID[option],dict5[option]))
+    else:
+        dict10={"P4":"Puissance 10","BatailleNavale":"Flotte Armée","Tortues":"Tortue de Sport","TortuesDuo":"Co-Pilote","TrivialVersus":"Médaillé","TrivialParty":"Bourré","TrivialBR":"Survivant"}
+        dict10ID={"P4":40,"BatailleNavale":46,"Tortues":42,"TortuesDuo":44,"TrivialVersus":34,"TrivialParty":36,"TrivialBR":38}
+        curseurUser.execute("INSERT INTO titresUser VALUES({0},'{1}',0)".format(dict10ID[option],dict10[option]))
+    connexionUser.commit()
 
-async def monthlyTitles(mois,annee,bot):
-    dictTitres={"P4":"des Puissants","BatailleNavale":"des Mers","Tortues":"des Tortues","TortuesDuo":"de la Co-Op","TrivialVersus":"des Questions","TrivialParty":"de la Fête","TrivialBR":"de la Survie"}
-    dictID={"P4":144,"BatailleNavale":147,"Tortues":141,"TortuesDuo":150,"TrivialVersus":156,"TrivialParty":159,"TrivialBR":153}
-    liste=["P4","BatailleNavale","Tortues","TortuesDuo","TrivialVersus","TrivialParty","TrivialBR"]
-    for i in liste:
-        print(i)
-        connexion,curseur=connectSQL("OT",i,"Jeux",mois,annee)
-        try:
-            for j in curseur.execute("SELECT * FROM {0}{1} WHERE Rank=1".format(tableauMois[mois].lower(),annee)).fetchall():
-                connexionUser,curseurUser=connectSQL("OT",j["ID"],"Titres",None,None)
-                try:
-                    curseurUser.execute("INSERT INTO titresUser VALUES({0},'Prince {1}',0)".format(dictID[i],dictTitres[i]))
-                    user=bot.get_user(j["ID"])
-                    await user.send(embed=createEmbed("Bravo !","Vous avez terminé **1er** du classement du **{0}** le mois dernier !\nVous avez donc droit au titre exclusif **Prince {1}** !\nVous pouvez l'équiper avec la commande *OT!titre set {2}*, et consulter votre liste de titres avec *OT!titre perso*.".format(i,dictTitres[i],dictID[i]),0xf58d1d,"titre",user))
-                except:
-                    pass
-                connexionUser.commit()
-        except sqlite3.OperationalError:
-            pass
-
-        connexion,curseur=connectSQL("OT",i,"Jeux","GL","")
-        connexionTitre,curseurTitre=connectSQL("OT","Titres","Titres",None,None)
-        rank1=curseur.execute("SELECT * FROM glob WHERE Rank=1").fetchone()
-        old=curseurTitre.execute("SELECT * FROM monthly WHERE Jeu='{0}'".format(i)).fetchone()
-        if rank1["ID"]!=old["ID"]:
-            print(rank1["ID"])
-            connexionUser,curseurUser=connectSQL("OT",rank1["ID"],"Titres",None,None)
-            createAccount(connexionUser,curseurUser)
-            curseurUser.execute("INSERT INTO titresUser VALUES({0},'L’Empereur {1}',0)".format(dictID[i]+2,dictTitres[i]))
-            connexionUser.commit()
-
-            connexionUser,curseurUser=connectSQL("OT",old["ID"],"Titres",None,None)
-            createAccount(connexionUser,curseurUser)
-            curseurUser.execute("DELETE FROM titresUser WHERE ID={0}".format(dictID[i]+2))
-            if curseurTitre.execute("SELECT * FROM active WHERE MembreID={0}".format(old["ID"])).fetchone()!=None:
-                if curseurTitre.execute("SELECT * FROM active WHERE MembreID={0}".format(old["ID"])).fetchone()["TitreID"]==dictID[i]+2:
-                    curseurTitre.execute("UPDATE active SET TitreID=8 WHERE MembreID={0}".format(old["ID"]))
-            connexionUser.commit()
-
-            curseurTitre.execute("UPDATE monthly SET ID={0} WHERE Jeu='{1}'".format(rank1["ID"],i))
-            connexionTitre.commit()
-
-            try:
-                user=bot.get_user(rank1["ID"])
-                #await user.send(embed=createEmbed("Bravo !","Vous êtes actuellement **1er** du classement GÉNÉRAL du **{0}** !!\nVous avez donc droit au titre exclusif **Empereur {1}** pour un mois !\nDéfendez votre place pendant tout le mois pour le garder !\nVous pouvez l'équiper avec la commande *OT!titre set {2}*, et consulter votre liste de titres avec *OT!titre perso*.".format(i,dictTitres[i],dictID[i]),0xf58d1d,user))
-            except:
-                pass
+def getTitre(curseur,user):
+    total=""
+    custom=curseur.execute("SELECT Custom FROM custom WHERE ID={0}".format(user)).fetchone()
+    titre=curseur.execute("SELECT titres.Nom FROM active JOIN titres ON active.TitreID=titres.ID WHERE MembreID={0}".format(user)).fetchone()
+    if custom!=None:
+        total+="{0}, ".format(custom["Custom"])
+    if titre==None:
+        total+="Inconnu"
+    else:
+        total+=titre["Nom"]
+    return total
